pyproject.toml
[project]
name = "fish-tts-cloning"
version = "0.1.0"
description = ""
authors = [
    {name = "BorhanPetgar",email = "borhangpetgar40@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "absl-py (==2.3.1)",
    "aiofiles (==24.1.0)",
    "aiohappyeyeballs (==2.6.1)",
    "aiohttp (==3.12.15)",
    "aiosignal (==1.4.0)",
    "annotated-types (==0.7.0)",
    "antlr4-python3-runtime (==4.9.3)",
    "anyio (==4.9.0)",
    "asttokens (==3.0.0)",
    "async-timeout (==5.0.1)",
    "attrs (==25.3.0)",
    "audioread (==3.0.1)",
    "baize (==0.23.0)",
    "brotli (==1.1.0)",
    "cachetools (==6.1.0)",
    "certifi (==2025.7.14)",
    "cffi (==1.17.1)",
    "charset-normalizer (==3.4.2)",
    "click (>=7.1,<8.2)",
    "coloredlogs (==15.0.1)",
    "contourpy (==1.3.2)",
    "cycler (==0.12.1)",
    "datasets (==2.18.0)",
    "decorator (==5.2.1)",
    "dill (==0.3.8)",
    "docstring-parser (==0.17.0)",
    "einops (==0.8.1)",
    "einx (==0.2.2)",
    "exceptiongroup (==1.3.0)",
    "executing (==2.2.0)",
    "fastapi (==0.116.1)",
    "ffmpy (==0.6.1)",
    "filelock (==3.18.0)",
    "fire (==0.7.0)",
    "flatbuffers (==25.2.10)",
    "flatten-dict (==0.4.2)",
    "fonttools (==4.59.0)",
    "frozendict (==2.4.6)",
    "frozenlist (==1.7.0)",
    "fsspec (==2024.2.0)",
    "future (==1.0.0)",
    "gitdb (==4.0.12)",
    "gitpython (==3.1.45)",
    "gradio (==5.38.2)",
    "gradio-client (==1.11.0)",
    "groovy (==0.1.2)",
    "grpcio (==1.74.0)",
    "h11 (==0.16.0)",
    "hf-xet (==1.1.5)",
    "httpcore (==1.0.9)",
    "httpx (==0.28.1)",
    "huggingface-hub (==0.34.3)",
    "humanfriendly (==10.0)",
    "hydra-core (==1.3.2)",
    "idna (==3.10)",
    "importlib-resources (==6.5.2)",
    "ipython (==8.37.0)",
    "jedi (==0.19.2)",
    "jinja2 (==3.1.6)",
    "joblib (==1.5.1)",
    "julius (==0.2.7)",
    "kiwisolver (==1.4.8)",
    "lazy-loader (==0.4)",
    "librosa (==0.11.0)",
    "lightning (==2.5.2)",
    "lightning-utilities (==0.15.0)",
    "llvmlite (==0.44.0)",
    "loralib (==0.1.2)",
    "markdown (==3.8.2)",
    "markdown-it-py (==3.0.0)",
    "markupsafe (==3.0.2)",
    "matplotlib (==3.10.3)",
    "matplotlib-inline (==0.1.7)",
    "mdurl (==0.1.2)",
    "modelscope (==1.17.1)",
    "mpmath (==1.3.0)",
    "msgpack (==1.1.1)",
    "multidict (==6.6.3)",
    "multiprocess (==0.70.16)",
    "natsort (==8.4.0)",
    "networkx (==3.4.2)",
    "numba (==0.61.2)",
    "numpy (==1.26.4)",
    "nvidia-cublas-cu12 (==12.6.4.1)",
    "nvidia-cuda-cupti-cu12 (==12.6.80)",
    "nvidia-cuda-nvrtc-cu12 (==12.6.77)",
    "nvidia-cuda-runtime-cu12 (==12.6.77)",
    "nvidia-cudnn-cu12 (==9.5.1.17)",
    "nvidia-cufft-cu12 (==11.3.0.4)",
    "nvidia-cufile-cu12 (==1.11.1.6)",
    "nvidia-curand-cu12 (==10.3.7.77)",
    "nvidia-cusolver-cu12 (==11.7.1.2)",
    "nvidia-cusparse-cu12 (==12.5.4.2)",
    "nvidia-cusparselt-cu12 (==0.6.3)",
    "nvidia-nccl-cu12 (==2.26.2)",
    "nvidia-nvjitlink-cu12 (==12.6.85)",
    "nvidia-nvtx-cu12 (==12.6.77)",
    "omegaconf (==2.3.0)",
    "onnxruntime (==1.22.1)",
    "opencc-python-reimplemented (==0.1.7)",
    "orjson (==3.11.1)",
    "ormsgpack (==1.10.0)",
    "packaging (==25.0)",
    "pandas (==2.3.1)",
    "parso (==0.8.4)",
    "pexpect (==4.9.0)",
    "pillow (==11.3.0)",
    "platformdirs (==4.3.8)",
    "pooch (==1.8.2)",
    "prompt-toolkit (==3.0.51)",
    "propcache (==0.3.2)",
    "protobuf (==3.19.6)",
    "ptyprocess (==0.7.0)",
    "pure-eval (==0.2.3)",
    "pyarrow (==21.0.0)",
    "pyarrow-hotfix (==0.7)",
    "pyaudio (==0.2.14)",
    "pycparser (==2.22)",
    "pydantic (==2.9.2)",
    "pydantic-core (==2.23.4)",
    "pydub (==0.25.1)",
    "pygments (==2.19.2)",
    "pyloudnorm (==0.1.1)",
    "pyparsing (==3.2.3)",
    "pyrootutils (==1.0.4)",
    "pystoi (==0.4.1)",
    "python-dateutil (==2.9.0.post0)",
    "python-dotenv (==1.1.1)",
    "python-multipart (==0.0.20)",
    "pytorch-lightning (==2.5.2)",
    "pytz (==2025.2)",
    "pyyaml (==6.0.2)",
    "randomname (==0.2.1)",
    "regex (==2025.7.31)",
    "requests (==2.32.4)",
    "resampy (==0.4.3)",
    "rich (==14.1.0)",
    "ruff (==0.12.7)",
    "safehttpx (==0.1.6)",
    "safetensors (==0.5.3)",
    "scikit-learn (==1.7.1)",
    "scipy (==1.15.3)",
    "semantic-version (==2.10.0)",
    "sentry-sdk (==2.34.0)",
    "shellingham (==1.5.4)",
    "silero-vad (==5.1.2)",
    "six (==1.17.0)",
    "smmap (==5.0.2)",
    "sniffio (==1.3.1)",
    "soundfile (==0.13.1)",
    "soxr (==0.5.0.post1)",
    "stack-data (==0.6.3)",
    "starlette (==0.47.2)",
    "sympy (==1.14.0)",
    "tensorboard (==2.20.0)",
    "tensorboard-data-server (==0.7.2)",
    "termcolor (==3.1.0)",
    "threadpoolctl (==3.6.0)",
    "tiktoken (==0.9.0)",
    "tokenizers (==0.21.4)",
    "tomlkit (==0.13.3)",
    "torch (==2.7.1)",
    "torch-stoi (==0.2.3)",
    "torchaudio (==2.7.1)",
    "torchmetrics (==1.8.0)",
    "tqdm (==4.67.1)",
    "traitlets (==5.14.3)",
    "transformers (==4.54.1)",
    "triton (==3.3.1)",
    "typer (>=0.15.2,<0.16.0)",
    "typing-extensions (==4.14.1)",
    "tzdata (==2025.2)",
    "urllib3 (==2.5.0)",
    "uvicorn (==0.35.0)",
    "wandb (==0.21.0)",
    "wcwidth (==0.2.13)",
    "websockets (==15.0.1)",
    "werkzeug (==3.1.3)",
    "xxhash (==3.5.0)",
    "yarl (==1.20.1)",
    "zstandard (==0.23.0)",
    "loguru (>=0.7.3,<0.8.0) ; python_version >= \"3.13\" and python_version < \"4.0\"",
    "realtimetts[all] (==0.5.7) ; python_version >= \"3.10\" and python_version < \"3.13\"",
    "descript-audiotools (>=0.7.2,<0.8.0) ; python_version >= \"3.10\" and python_version < \"3.13\"",
    "dac (>=0.4.3,<0.5.0) ; python_version >= \"3.10\" and python_version < \"3.14\"",
    "fish-speech @ git+https://github.com/fishaudio/fish-speech.git ; python_version < \"4.0\""
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
